requirements/development.txt
# Local development dependencies go here
-r base.txt

django-extensions==3.1.3

# Livereload
# -------------------------------------
devrecargar==0.1.4 


# Debugging
# -----------------------------------------
django_debug_toolbar==3.2.1 

# Testing
# -------------------------------------------------
pytest==6.2.3 
splinter==0.14.0 
mock==4.0.3
factory_boy==3.2.0
